Use named functions for stateless functional components.
That way they show up named properly in React devtools.
