+++ src/components/form/form.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+function Form({ children, handleSubmit }) {
+  return (
+    <form
+      onSubmit={(e) => {
+        e.preventDefault();
+        document.activeElement.blur();
+        handleSubmit();
+      }}>
+      { children }
+    </form>
+  );
+}
+
+Form.propTypes = {
+  children: React.PropTypes.node,
+  handleSubmit: React.PropTypes.func,
+};
+
+export default Form;
