--- src/components/login-modal.js
@@ -1,26 +1,0 @@
-import React from 'react';
-
-import {Modal, ModalContent} from './modal';
-import LoginForm from './login-form';
-
-const LoginModal = ({ isVisible, isPending, hasError, onSubmit }) => (
-  <Modal isVisible={ isVisible }>
-    <ModalContent>
-      <h1 className="mt0">Login</h1>
-
-      <LoginForm
-        isPending={ isPending }
-        hasError={ hasError }
-        onSubmit={ onSubmit } />
-    </ModalContent>
-  </Modal>
-);
-
-LoginModal.propTypes = {
-  isVisible: React.PropTypes.bool,
-  isPending: React.PropTypes.bool,
-  hasError: React.PropTypes.bool,
-  onSubmit: React.PropTypes.func,
-};
-
-export default LoginModal;
