+++ src/components/login/login-modal.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+import { Modal, ModalContent } from '../modal';
+import LoginForm from './login-form';
+
+function LoginModal({ isVisible, isPending, hasError, onSubmit }) {
+  return (
+    <Modal isVisible={ isVisible }>
+      <ModalContent>
+        <h1 className="mt0">Login</h1>
+
+        <LoginForm
+          isPending={ isPending }
+          hasError={ hasError }
+          onSubmit={ onSubmit } />
+      </ModalContent>
+    </Modal>
+  );
+}
+
+LoginModal.propTypes = {
+  isVisible: React.PropTypes.bool,
+  isPending: React.PropTypes.bool,
+  hasError: React.PropTypes.bool,
+  onSubmit: React.PropTypes.func,
+};
+
+export default LoginModal;
