+++ src/components/logo.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import LogoImage from '../assets/rangleio-logo.svg';
 
-const Logo = () => (
-  <div className="flex items-center">
-    <img style={ styles }
-      src={ LogoImage }
-      alt="Rangle.io" />
-  </div>
-);
+function Logo() {
+  return (
+    <div className="flex items-center">
+      <img style={ styles }
+        src={ LogoImage }
+        alt="Rangle.io" />
+    </div>
+  );
+}
 
 const styles = { width: 128 };
