+++ src/components/modal/modal-content.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const ModalContent = ({ children, close }) => {
+function ModalContent({ children, close }) {
   return (
     <div className="p2 z2 bg-white modal relative">
       <button className="btn absolute top-0 right-0 m1"
@@ -10,7 +10,7 @@ const ModalContent = ({ children, close }) => {
       { children }
     </div>
   );
-};
+}
 
 ModalContent.propTypes = {
   children: React.PropTypes.node,
