+++ src/components/navigator.js
@@ -1,10 +1,12 @@
 import React from 'react';
 
-const Navigator = ({ children }) => (
-  <nav className="flex items-center p1 bg-white border-bottom">
-    { children }
-  </nav>
-);
+function Navigator({ children }) {
+  return (
+    <nav className="flex items-center p1 bg-white border-bottom">
+      { children }
+    </nav>
+  );
+}
 
 Navigator.propTypes = {
   children: React.PropTypes.node,
